+++ aiohttp_tus/__init__.py
@@ -1,3 +1,5 @@
 from .tus import setup_tus
 
 __all__ = ("setup_tus",)
+__license__ = "BSD-3-Clause"
+__version__ = "1.0.0a1"